% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy.R
\name{create_dummy}
\alias{create_dummy}
\title{Title}
\usage{
create_dummy(name = "newdummy", symptoms, data = vaers_2021)
}
\arguments{
\item{name}{The name the new variable should have}

\item{symptoms}{A vector of strings (symptoms) for which the function should search}

\item{data}{A dataset from the vaers datasets in this package}
}
\value{
Returns the dataset with a new variable
}
\description{
Title
}
\examples{
create_dummy("breathlessness", c("Breathlessness"), vaers_2018))
}
