% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_likelihood.R
\name{max_likelihood}
\alias{max_likelihood}
\title{Maximum likelihood parameter estimation}
\usage{
max_likelihood(.data)
}
\arguments{
\item{.data}{A data vector}
}
\value{

}
\description{
Maximum likelihood parameter estimation
}
