% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_vaccines.R
\name{show_vaccines}
\alias{show_vaccines}
\title{Shows vaccine names}
\usage{
show_vaccines(data = vaers_2020)
}
\arguments{
\item{data}{}
}
\value{
A sorted table of vaccines names
}
\description{
Shows vaccine names
}
\examples{
show_vaccines() # default is 2020
show_vaccines(data = vaers_2019)
vaers_2017 \%>\% show_vaccines()
}
