% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Title}
\usage{
create_dataset(.data = vaers_2020, vaxname)
}
\arguments{
\item{vaxname}{Name of the vaccine you want to filter for}

\item{data}{One of the vaers datasets from this package}
}
\value{
Returns the filtered dataset with a new column order.
}
\description{
Title
}
\examples{
create_dataset(vaers_2021, "COVID19 (COVID19 (JANSSEN))")
}
