% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummy.R
\name{create_dummy}
\alias{create_dummy}
\title{Title}
\usage{
create_dummy(.data = vaers_2020, name = "newdummy", symptoms)
}
\arguments{
\item{.data}{A dataset from the vaers datasets in this package (named "vaers_20xx")}

\item{name}{The name the new variable should have}

\item{symptoms}{A vector of strings (symptoms) for which the function should search}
}
\value{
Returns the input dataset with a new dummy-variable attached
}
\description{
Title
}
\examples{
create_dummy(vaers_2018, "breathlessness", c("Breathlessness"))
vaers_2020 \%>\% create_dummy("fatigue", c("Fatigue", "Lethargy", "Somnolence", "Asthenia", "Hypersomnia", "Depressed level of consciousness"))
}
